Made permanent proxy buffer for private Metal access that scales according to need, implemented pimpl to separate objc++, fleshed out runtime a bit more, and a few other api changes

+++ inc/IGPUExecutor.h
@@ -1,6 +1,7 @@
 #ifndef IGPU_H
 #define IGPU_H
 
+#include <cstddef>
 #include <functional>
 #include <span>
 #include <unordered_map>
@@ -12,7 +13,9 @@ enum class GPUState { GPUSuccess, GPUFailure, GhostBuffer, InvalidDispatchType }
 // BufferUsage - How the buffer will be accessed during GPU tasks, configurable so buffers can be reused
 // MemoryHint - How the data stored in the buffer will be treated throughout the lifetime of a task on a CPU/GPU level
 //  - Enables optimizations with private memory on the GPU, guarantee that only it has access to it
+
 // NOTE: Should BufferUsage be a more generalized data value? Might have some value on the CPU/other operations as well?
+// Buffer Usage used for Scheduling purposes primarily (e.g. two kernels can read from a buffer at the same time)
 enum class BufferUsage { ReadWrite, ReadOnly };
 enum class MemoryHint { DeviceLocal, HostVisible };
 
@@ -66,14 +69,14 @@ enum class DispatchType { Serial, Concurrent };
 class IGPUExecutor {
   public:
     // Allocating/freeing buffer memory on the GPU for kernel tasks
-    GPUBufferHandle virtual allocate_buffer(std::uint32_t buffer_size, const MemoryHint &mem_hint) = 0;
+    GPUBufferHandle virtual allocate_buffer(std::size_t buffer_size, const MemoryHint &mem_hint) = 0;
     GPUState virtual deallocate_buffer(const GPUBufferHandle &buffer_handle) = 0;
 
     // Sending memory between devices for task completion/after task completion
     GPUState virtual copy_to_device(std::span<const std::byte> data_mem, const GPUBufferHandle &buffer_handle,
-                                    std::uint32_t data_size) = 0;
+                                    std::size_t data_size) = 0;
     GPUState virtual copy_from_device(std::span<std::byte> data_mem, const GPUBufferHandle &buffer_handle,
-                                      std::uint32_t data_size, bool sync) = 0;
+                                      std::size_t data_size, bool sync) = 0;
 
     GPUState virtual execute_batch(const std::vector<KernelDispatch> &kernels, const DispatchType &dispatch_type,
                                    std::function<void()> &cpu_callback) = 0;
